Refactor to externalize route expansion
Rather than making the HyperspaceNetwork responsible for managing how
many routes should be discovered, we've simplified its interface to just
a method that discovers a single route. We've also relocated the growing
list of parameters from that method to the `__init__` method, which
reduces the potential for error and simplifies the interfaces of
internal methods.