Fixing projection errors
Added a few new reference stars and, using these points, identified that
I was unnecessarily flipping the y-axis, and needed to instead flip the
z-axis.

I'm frankly still not sure why I needed to flip the z-axis, that doesn't
make sense to me, but this is what I needed to do to get the results I
wanted (positive z pointing "up" in the final image).