Don't let route flares overwrite other routes
Previously, because we were drawing routes and "flares" on the same
"layer" of the image, the color/alpha of a flare was replacing anything
previously drawn, resulting in some unseemly visual "pops". We now draw
flares on their own, separate layer, and blend them together with the
rest of the image appropriately.